pyproject.toml
[tool.poetry]
name = "atlas-forward-proton-anomaly-detector"
version = "0.1.0"
description = ""
authors = ["Janusz Twardak"]

[tool.poetry.dependencies]
python = "^3.8"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
notebook = "^6.4.10"
pre-commit = "^2.19.0"
pandas = "^1.5.1"
numpy = "^1.23.4"
scipy = "^1.9.3"
matplotlib = "^3.6.2"
scikit-learn = "^1.1.3"
jupyter = "^1.0.0"
ipykernel = "^6.17.1"
seaborn = "^0.12.1"
uproot = "^4.3.7"
awkward = "^1.10.2"
pathlib2 = "^2.3.7"
dask = {extras = ["complete"], version = "^2022.11.0"}
Cython = "^0.29.32"
pyarrow = "^10.0.0"
p5py = "^1.0.0"
pep517 = "^0.13.0"
tqdm = "^4.64.1"
isotree = "^0.5.17"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''
